requirements.txt
fastapi
uvicorn
openai
beautifulsoup4
python-multipart
python-dotenv
selenium
webdriver-manager

